Custom Video Player part1

+++ controllers/videoController.js
@@ -5,7 +5,6 @@ import Video from "../models/Video";
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({}).sort({ _id: -1 });
-    console.log(routes.editProfile);
     res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
     console.log(error);
@@ -33,7 +32,6 @@ export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
 
 export const postUpload = async (req, res) => {
-  console.log(req);
   const {
     body: { title, description },
     file: { path }
@@ -44,20 +42,18 @@ export const postUpload = async (req, res) => {
     description,
     creator: req.user.id
   });
-  // req.user.videos.push(newVideo.id);
-  // req.user.save();
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   // console.log(newVideo);
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
 export const videoDetail = async (req, res) => {
-  console.log(req.params);
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id).populated("creator");
-    console.log(video);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -70,7 +66,13 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("videoEdit", { pageTitle: `Edt ${video.title}`, video });
+    console.log(typeof video.creator);
+    console.log(typeof req.user.id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("videoEdit", { pageTitle: `Edt ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -94,7 +96,14 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
-  } catch (error) {}
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
+  } catch (error) {
+    console.log(error);
+  }
   res.redirect(routes.home);
 };
